+++ src/components/LogOutBtn/logoutbtn.css
@@ -1,0 +1,10 @@
+.logOutbtn {
+    position: absolute;
+    left: 0;
+    top: 10%;
+    margin: 35px;
+    padding: 10px;
+    color: #fff;
+    background: firebrick;
+    border-radius: 7px;
+}
